When I was investigating https://github.com/rectorphp/rector/issues/8793 yesterday I noticed that the `NameImportingPostRector` post rector could be optimised because it was unnecessarily calculating the use statements for every node so I decided to add a fix. I took the opportunity to look at the other post rectors and found some other possible optimisations. They are small optimisations, nothing really very noticeable, but, as they say, every little helps!

See the PR comments for more info