@TomasVotruba 2024-08-20
Thanks for the improvement :+1: 

Could you share few time measurements, as on the other PR?