@carlos-granados 2024-08-20
@TomasVotruba time improvements are not really noticeable, I see the same timings more or less with the previous version and this one (the results vary a little but they vary equally in both versions). It is more a question of trying to be optimal overall so that all these little increments don't add up