@TomasVotruba 2024-08-20
We're adding a state to readonly service here, which I don't see as a good practise. 
I'd prefer to have measurable improvement, to balance the added complexity. We've added few "optimizations" in the past, but they added more complexity than real value so they were removed later.