@carlos-granados 2024-08-21
I have tested this without using parallel, which gives more stable results that don't vary too much and I can see numbers similar to these:
Without the optimizations:
`bin/rector --clear-cache  97.41s user 2.47s system 98% cpu 1:41.25 total`
With the optimizations:
`bin/rector --clear-cache  97.26s user 2.41s system 98% cpu 1:41.06 total`
This represents an improvement of around 0.3%, so not very significant.
I would say that we should probably keep the changes to `UnusedImportRemovingPostRector` and `UseAddingPostRector` as they don't really add more complexity and I'll let you decide about the other two