@TomasVotruba 2024-08-22
Thanks for checking :+1: 

> I would say that we should probably keep the changes to UnusedImportRemovingPostRector and UseAddingPostRector as they don't really add more complexity and I'll let you decide about the other two

You're right. Let's ship this. Could you rebase the PR?